requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python==4.8.1.78
spacy==3.7.2
mysql-connector-python==8.1.0
pandas==2.1.3
streamlit==1.28.1
python-dotenv==1.0.0
pdf2image==1.16.3
pillow==10.1.0
regex==2023.10.3
plotly==5.18.0

# Download spaCy model
# python -m spacy download en_core_web_sm